+++ src/compiler/constants.rs
@@ -59,7 +59,7 @@ pub(crate) const JSON_FILE_EXTENSION: &str = "json";
 
 // findings
 pub(crate) const FINDINGS_DIR_NAME: &str = "mend-findings";
-pub(crate) const FINDINGS_SCHEMA_VERSION: u32 = 16;
+pub(crate) const FINDINGS_SCHEMA_VERSION: u32 = 17;
 
 // progress indicator
 pub(super) const PROGRESS_FRAMES: [&str; 4] = ["|", "/", "-", "\\"];